files for design problem 2
project description is as follows:

It is the year 2020, and the zombie apocalypse has been terrorizing the world for 2 years now. You are the computer scientist in a team of scientists trying to identify the different types of zombies and their weaknesses. You want to be able to search for a type of zombie, and have the program return that zombie’s weakness. You also want to be able to do the search query as quickly as possible, as there are a constantly-growing number of zombie types.